pyproject.toml
[tool.poetry]
name = "solar-works-enerflo"
version = "0.1.0"
description = ""
authors = ["HM <hieumdd@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9"
requests = "^2.27.1"
google-cloud-bigquery = "^3.0.1"
compose = "^1.2.8"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.1"
mypy = "^0.942"
types-requests = "^2.27.16"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
